Added possible fix
By multiplying rateHave and rateNeed's values by 1.0, it forces the program to use double/float/etc. numbers instead of just integers, so it may prevent the program from consistently reporting huge whole numbers that jump back and forth because of that type of math; instead it'll report more accurate numbers. There will be another commit if this approach doesn't work~